+++ src/main/java/org/eclipse/tractusx/autosetup/constant/SDEConfigurationProperty.java
@@ -1,6 +1,6 @@
 /********************************************************************************
-#* Copyright (c) 2022, 2023 T-Systems International GmbH
-#* Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+#* Copyright (c) 2022,2024 T-Systems International GmbH
+#* Copyright (c) 2022,2024 Contributors to the Eclipse Foundation
 #*
 #* See the NOTICE file(s) distributed with this work for additional
 #* information regarding copyright ownership.
@@ -57,5 +57,9 @@ public class SDEConfigurationProperty {
 	private String discoveryAuthenticationUrl;
 	private String discoveryClientId;
 	private String discoveryClientSecret;
+	
+	private String bpdmProviderEdcDataspaceApi;
+	private String bpdmProviderBpnl;
+	private String bpdmProviderEdcPublicApi;
 	
 }