+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+
+### Added
+- Support Portal autosetup latest workflow.
+- Support SDE, Digital twin, EDC 7 support 24/05 release.
+
 ### Fixed
 - Dependabot reported security issues fixed.